Create numbers_searching
Write a function called numbers_searching which receives a different amount of parameters. All parameters will be integer numbers forming a sequence of consecutive numbers. Your task is to find an unknown amount of duplicates from the given sequence and a missing value, such that all the duplicate values and the missing value are between the smallest and the biggest received number.

The function should return a list with the last missing number as a first argument and a sorted list, containing the duplicates found, in ascending order.